refactor(constant): 상수 용도에 맞게 파일로 분리
