+++ src/constant/lotto.js
@@ -1,0 +1,12 @@
+export const LOTTO_MONEY = 1000;
+export const LOTTO_MAX_MONEY = 100000;
+export const LOTTO_MIN_NUMBER = 1;
+export const LOTTO_MAX_NUMBER = 45;
+export const LOTTO_LENGTH = 6;
+export const PRIZE_MAP = {
+	"5rank": 5000,
+	"4rank": 50000,
+	"3rank": 150000,
+	"2rank": 30000000,
+	"1rank": 2000000000,
+};
